refactor: capitalize component files
